# openapi/artifacts/assistant-api.yaml
openapi: 3.0.3
info:
  title: Rapida Assistant API
  version: 0.1.0
paths:
...
  /v1/assistant-deployment/get-api-deployment/{assistantId}:
    get:
      operationId: getAssistantApiDeployment
      summary: Get assistant API deployment
      description: Gets the latest API deployment for an assistant. This endpoint requires user authentication with organization and project scope.
      parameters:
        - name: assistantId
          in: path
          required: true
          schema:
            $ref: "./common.yaml#/components/schemas/Uint64String"
      responses:
        "200":
          description: Assistant API deployment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAssistantApiDeploymentResponse"
        "400":
          description: Invalid assistantId
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/ErrorResponse"
        "401":
          description: Unauthenticated request
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/ErrorResponse"
        "403":
          description: Missing authentication scope
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/ErrorResponse"
        "500":
          description: Failed to get assistant API deployment
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/ErrorResponse"
  /v1/assistant-deployment/get-debugger-deployment/{assistantId}:
    get:
      operationId: getAssistantDebuggerDeployment
      summary: Get assistant debugger deployment
      description: Gets the latest debugger deployment for an assistant. This endpoint requires user authentication with organization and project scope.
      parameters:
        - name: assistantId
          in: path
          required: true
          schema:
            $ref: "./common.yaml#/components/schemas/Uint64String"
      responses:
        "200":
          description: Assistant debugger deployment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAssistantDebuggerDeploymentResponse"
        "400":
          description: Invalid assistantId
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/ErrorResponse"
        "401":
          description: Unauthenticated request
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/ErrorResponse"
        "403":
          description: Missing authentication scope
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/ErrorResponse"
        "500":
          description: Failed to get assistant debugger deployment
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    GetAssistantDebuggerDeploymentResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/AssistantDebuggerDeployment"
        error:
          $ref: "./common.yaml#/components/schemas/Error"
...
    GetAssistantApiDeploymentResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/AssistantApiDeployment"
        error:
          $ref: "./common.yaml#/components/schemas/Error"
...
    AssistantDebuggerDeployment:
      type: object
      properties:
        id:
          $ref: "./common.yaml#/components/schemas/Uint64String"
        assistantId:
          $ref: "./common.yaml#/components/schemas/Uint64String"
        greeting:
          type: string
        greetingInterruptible:
          type: boolean
        mistake:
          type: string
        inputAudio:
          $ref: "#/components/schemas/DeploymentAudioProvider"
        outputAudio:
          $ref: "#/components/schemas/DeploymentAudioProvider"
        status:
          type: string
        maxSessionDuration:
          type: integer
          format: uint64
          minimum: 180
          maximum: 600
        idealTimeout:
          type: integer
          format: uint64
          minimum: 15
          maximum: 120
        idealTimeoutBackoff:
          type: integer
          format: uint64
          minimum: 0
          maximum: 5
        idealTimeoutMessage:
          type: string
...
    AssistantApiDeployment:
      type: object
      properties:
        id:
          $ref: "./common.yaml#/components/schemas/Uint64String"
        assistantId:
          $ref: "./common.yaml#/components/schemas/Uint64String"
        greeting:
          type: string
        greetingInterruptible:
          type: boolean
        mistake:
          type: string
        inputAudio:
          $ref: "#/components/schemas/DeploymentAudioProvider"
        outputAudio:
          $ref: "#/components/schemas/DeploymentAudioProvider"
        status:
          type: string
        maxSessionDuration:
          type: integer
          format: uint64
          minimum: 180
          maximum: 600
        idealTimeout:
          type: integer
          format: uint64
          minimum: 15
          maximum: 120
        idealTimeoutBackoff:
          type: integer
          format: uint64
          minimum: 0
          maximum: 5
        idealTimeoutMessage:
          type: string
...
    DeploymentAudioProvider:
      type: object
      properties:
        id:
          $ref: "./common.yaml#/components/schemas/Uint64String"
        audioType:
          type: string
        audioProvider:
          type: string
        audioOptions:
          type: array
          items:
            $ref: "./common.yaml#/components/schemas/Metadata"
        status:
          type: string
...